requirements.txt
fastapi==0.115.3
pydantic==2.9.2
pydantic_settings==2.6.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.0
aiosqlite
httpx[socks]
moviepy
imageio
apscheduler
aiohttp